+++ app/Http/Middleware/Development.php
@@ -11,18 +11,27 @@ use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 class Development
 {
+    const DEVELOPMENT_MODE_UP_TIME_SEC = 60 * 10;
+
     /**
      * Handle an incoming request.
      *
-     * @param  \Closure(\Illuminate\Http\Request): (\Symfony\Component\HttpFoundation\Response)  $next
+     * @param \Closure(\Illuminate\Http\Request): (\Symfony\Component\HttpFoundation\Response) $next
      *
      * @throws NotFoundHttpException
      */
     public function handle(Request $request, Closure $next): Response
     {
-        if (Configuration::getBool(ConfigurationKey::DEVELOPMENT_PAGE_AVAILABLE, default: false)) {
+        if (self::isAvailableInSec() > 0) {
             return $next($request);
         }
         throw new NotFoundHttpException();
+    }
+
+    public static function isAvailableInSec(): int
+    {
+        return Configuration::getInt(ConfigurationKey::DEVELOPMENT_PAGE_AVAILABLE, default: 0)
+            + self::DEVELOPMENT_MODE_UP_TIME_SEC
+            - now()->getTimestamp();
     }
 }